requirements.txt
aniso8601==8.0.0
Click==7.0
Flask==1.1.1
flask-marshmallow==0.10.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
marshmallow==3.2.2
pytz==2019.3
six==1.13.0
SQLAlchemy==1.3.10
Werkzeug==0.16.0
flask_marshmallow
yfinance
flask-apscheduler
Gunicorn
psycopg2-binary
marshmallow-sqlalchemy=0.22.3
